Move betty.plugins.loader to betty.loaders
