+++ betty/demo/project.py
@@ -12,6 +12,7 @@ from betty.extension import ExtensionManufacturer
 from betty.license import LicenseManufacturer
 from betty.links.betty_documentation import BETTY_DOCUMENTATION
 from betty.links.betty_github import BETTY_GITHUB
+from betty.loaders.demo import Demo
 from betty.locale import DEFAULT_LOCALE
 from betty.locale.localizable.gettext import _
 from betty.locale.localizable.markup import Chain
@@ -39,7 +40,6 @@ from betty.plugins.extension.raspberry_mint import (
 )
 from betty.plugins.extension.spdx import Spdx
 from betty.plugins.extension.trees import Trees
-from betty.plugins.loader.demo import Demo
 from betty.plugins.media_type.html import HTML
 from betty.project import Project
