+++ betty/tests/loaders/demo/test___init__.py
@@ -5,12 +5,12 @@ from typing import TYPE_CHECKING
 import pytest
 
 from betty.load import load
+from betty.loaders.demo import Demo
 from betty.plugins.entity.citation import Citation
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.place import Place
 from betty.plugins.entity.source import Source
-from betty.plugins.loader.demo import Demo
 
 if TYPE_CHECKING:
     from pytest_mock import MockerFixture
