+++ betty/tests/loaders/demo/test_project.py
@@ -5,7 +5,7 @@ import pytest
 from betty.copyright_notice import CopyrightNotice
 from betty.entity.collection.pool import EntityPool
 from betty.license import License
-from betty.plugins.loader.demo import LoadAncestry
+from betty.loaders.demo import LoadAncestry
 from betty.service_level import ServiceLevel
 from betty.test_utils.job import do
