+++ betty/tests/loaders/gramps/test___init__.py
@@ -5,6 +5,7 @@ from tempfile import TemporaryDirectory
 from betty.event_type import EventTypeManufacturer
 from betty.event_types.birth import Birth
 from betty.load import LoaderManufacturer, load
+from betty.loaders.gramps import FamilyTree, Gramps, GrampsData
 from betty.place_type import PlaceTypeManufacturer
 from betty.place_types.borough import Borough
 from betty.place_types.city import City
@@ -14,7 +15,6 @@ from betty.plugins.entity.note import Note
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.place import Place
 from betty.plugins.entity.source import Source
-from betty.plugins.loader.gramps import FamilyTree, Gramps, GrampsData
 from betty.role import RoleManufacturer
 from betty.roles.attendee import Attendee
 from betty.roles.subject import Subject
